pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "umbra-tpp"
version = "0.1.0"
description = "Causal ML framework for hidden liquidity discovery using Neural Temporal Point Processes"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "econml>=0.15",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
